Fraction->double in SwAccessibleMap
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202402
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/202818
Tested-by: Tomaž Vajngerl <quikee@gmail.com>

+++ sw/source/core/view/viewsh.cxx
@@ -2748,7 +2748,7 @@ rtl::Reference<comphelper::OAccessible> SwViewShell::CreateAccessiblePreview()
     {
         return Imp()->GetAccessibleMap().GetDocumentPreview(
                     PagePreviewLayout()->maPreviewPages,
-                    GetWin()->GetMapMode().GetScaleX(),
+                    double(GetWin()->GetMapMode().GetScaleX()),
                     GetLayout()->GetPageByPageNum( PagePreviewLayout()->mnSelectedPageNum ),
                     PagePreviewLayout()->maWinSize );
     }
